+++ packages/dodo-ui-bits/src/lib/index.ts
@@ -38,24 +38,24 @@ export { type SelectInputProps } from './components/Form/Select/SelectInput.svel
 export { type SelectPopupProps } from './components/Form/Select/SelectPopup.svelte';
 
 // =====================================================
-// Modals Components – Public API
+// Feedback Components – Public API
 // =====================================================
 
 /**
  * Modal component and related prop types.
  */
-export { default as Modal, type ModalProps } from './components/Modals/Modal/Modal.svelte';
+export { default as Modal, type ModalProps } from './components/Feedback/Modals/Modal/Modal.svelte';
 
 export {
   default as ModalCard,
   type ModalCardProps,
-} from './components/Modals/Modal/ModalCard/ModalCard.svelte';
+} from './components/Feedback/Modals/Modal/ModalCard/ModalCard.svelte';
 
-export { type ModalContentControlsProps } from './components/Modals/Modal/ModalCard/Controls.svelte';
+export { type ModalContentControlsProps } from './components/Feedback/Modals/Modal/ModalCard/Controls.svelte';
 
-export { type ModalContentDescriptionProps } from './components/Modals/Modal/ModalCard/Description.svelte';
+export { type ModalContentDescriptionProps } from './components/Feedback/Modals/Modal/ModalCard/Description.svelte';
 
-export { type ModalContentTitleProps } from './components/Modals/Modal/ModalCard/Title.svelte';
+export { type ModalContentTitleProps } from './components/Feedback/Modals/Modal/ModalCard/Title.svelte';
 
 /**
  * InformDialog component and related prop types.
@@ -63,13 +63,13 @@ export { type ModalContentTitleProps } from './components/Modals/Modal/ModalCard
 export {
   default as InformDialog,
   type InformDialogProps,
-} from './components/Modals/InformDialog/InformDialog.svelte';
+} from './components/Feedback/Modals/InformDialog/InformDialog.svelte';
 
 export {
   default as InformDialogCard,
   type InformDialogCardProps,
   type ModalActionButtonProps,
-} from './components/Modals/InformDialog/InformDialogCard/InformDialogCard.svelte';
+} from './components/Feedback/Modals/InformDialog/InformDialogCard/InformDialogCard.svelte';
 
 /**
  * ConfirmDialog component and related prop types.
@@ -77,12 +77,12 @@ export {
 export {
   default as ConfirmDialog,
   type ConfirmDialogProps,
-} from './components/Modals/ConfirmDialog/ConfirmDialog.svelte';
+} from './components/Feedback/Modals/ConfirmDialog/ConfirmDialog.svelte';
 
 export {
   default as ConfirmDialogCard,
   type ConfirmDialogCardProps,
-} from './components/Modals/ConfirmDialog/ConfirmDialogCard/ConfirmDialogCard.svelte';
+} from './components/Feedback/Modals/ConfirmDialog/ConfirmDialogCard/ConfirmDialogCard.svelte';
 
 /**
  * ModalManager component and related prop types.
@@ -90,7 +90,7 @@ export {
 export {
   default as ModalManager,
   type ModalManagerProps,
-} from './components/Modals/ModalManager/ModalManager.svelte';
+} from './components/Feedback/Modals/ModalManager/ModalManager.svelte';
 
 export {
   modals,
@@ -100,4 +100,4 @@ export {
   type ModalDialogInformConfig,
   type ModalDialogConfirmConfig,
   type ActiveModal,
-} from './components/Modals/ModalManager/modals.svelte.ts';
+} from './components/Feedback/Modals/ModalManager/modals.svelte.ts';